Use play framework for json parsing and serialization.

+++ config/sql/sketchy.sketchy_items-schema.sql
@@ -2,7 +2,7 @@
 /*!40014 SET FOREIGN_KEY_CHECKS=0*/;
 
 CREATE TABLE `sketchy_items` (
-  `id` int(11) unsigned NOT NULL,
+  `id` bigint unsigned NOT NULL,
   `kind` char(16) NOT NULL,
   `created_at` datetime DEFAULT NULL,
   PRIMARY KEY (`id`,`kind`)
