+++ core/src/main/scala/com.soundcloud.sketchy/access/SketchyAccess.scala
@@ -17,7 +17,7 @@ import com.soundcloud.sketchy.events.{
 }
 
 
-case class TrustedUser(user_id: Int, reason: String, created_at: Date)
+case class TrustedUser(user_id: Long, reason: String, created_at: Date)
 
 /**
  * Reputation in sketchy db
@@ -29,7 +29,7 @@ class SketchyAccess(db: com.soundcloud.sketchy.util.Database) {
   private val sketchyScores = TableQuery[SketchyScores]
   private val trustedUsers = TableQuery[TrustedUsers]
 
-  def select(userId: Int, kind: String): Option[SketchyScore] =
+  def select(userId: Long, kind: String): Option[SketchyScore] =
     db.withFailover("select", false) {
       val q =
         sketchyScores.filter(score => score.user_id === userId && score.kind === kind)
@@ -54,17 +54,17 @@ class SketchyAccess(db: com.soundcloud.sketchy.util.Database) {
     (updateStatus > 0) && (!items || mergeItems(sig))
   }
 
-  def trusted(userId: Int): Boolean =
+  def trusted(userId: Long): Boolean =
     !db.withFailover("trusted", false) {
       trustedUsers.filter(user => user.user_id === userId).list
     }.getOrElse(List()).isEmpty
 
-  def sketchy(kind: String, id: Int): Boolean =
+  def sketchy(kind: String, id: Long): Boolean =
     db.withFailover("sketchy", false) {
       selectItem(kind, id)
     }.getOrElse(None).isDefined
 
-  def selectItem(kind: String, id: Int): Option[SketchyItem] =
+  def selectItem(kind: String, id: Long): Option[SketchyItem] =
     db.withFailover("selectItem", false) {
       sketchyItems.filter(item => item.kind === kind && item.id === id).list
     }.getOrElse(List()).headOption
@@ -79,12 +79,12 @@ class SketchyAccess(db: com.soundcloud.sketchy.util.Database) {
     ).foldLeft(true)(_&&_)
   }
 
-  private def insertItem(id: Int, kind: String, createdAt: Date): Boolean =
+  private def insertItem(id: Long, kind: String, createdAt: Date): Boolean =
     db.withFailover("insertItem", true, isQuiet = true) {
       sketchyItems.insert(SketchyItem(id, kind, createdAt))
     }.getOrElse(0) > 0
 
-  private def updateItem(id: Int, kind: String, newCreatedAt: Date): Boolean =
+  private def updateItem(id: Long, kind: String, newCreatedAt: Date): Boolean =
     db.withFailover("updateItem", true) {
       val q = for { item <- sketchyItems if item.id === id && item.kind === kind }
         yield item.createdAt
@@ -99,14 +99,14 @@ class SketchyAccess(db: com.soundcloud.sketchy.util.Database) {
 
 
   private class SketchyItems(tag: Tag) extends Table[SketchyItem](tag, "sketchy_items") {
-    def id = column[Int]("id", O.PrimaryKey)
+    def id = column[Long]("id", O.PrimaryKey)
     def kind = column[String]("kind", O.PrimaryKey)
     def createdAt = column[Date]("created_at")
     def * = (id, kind, createdAt) <> (SketchyItem.tupled, SketchyItem.unapply)
   }
 
   private class SketchyScores(tag: Tag) extends Table[SketchyScore](tag, "sketchy_scores") {
-    def user_id = column[Int]("user_id")
+    def user_id = column[Long]("user_id")
     def kind = column[String]("kind")
     def signals = column[Int]("signals")
     def state = column[Int]("state")
@@ -126,7 +126,7 @@ class SketchyAccess(db: com.soundcloud.sketchy.util.Database) {
   }
 
   private class TrustedUsers(tag: Tag) extends Table[TrustedUser](tag, "trusted_users") {
-    def user_id = column[Int]("user_id", O.PrimaryKey) // This is the primary key column
+    def user_id = column[Long]("user_id", O.PrimaryKey) // This is the primary key column
     def reason = column[String]("reason")
     def created_at = column[Date]("created_at")
     def * = (user_id, reason, created_at) <> (TrustedUser.tupled, TrustedUser.unapply)
